tuning on the comp bot, didn't finish -- still have mecanumfeedback because the right rear wheel needed fixing. continue to tune -- changed a lot of motors from leftFront to frontLeft and same thing on other motors too.
